// src/models/dat-of-week.ts
/**
 * Inspired by Google Business API
 * @see https://developers.google.com/my-business/reference/rest/v4/accounts.locations#Location.DayOfWeek
 */
export enum DayOfWeek {
  DAY_OF_WEEK_UNSPECIFIED = -1,
  MONDAY = 1,
  TUESDAY = 2,
  WEDNESDAY = 3,
  THURSDAY = 4,
  FRIDAY = 5,
  SATURDAY = 6,
  SUNDAY = 0,
}